refactor: `processDeath` 모듈화
Ref: #116, #12
